Add a thread id display mode.
The thread id is encoded with the address sent on the pipe, using up 10
bits of the 64-bit value.  The number of bits available for addresses is
now down to 43, and for time is now 19 bits.  Thread id and type values are
packaged together during the most frequent update operations, to avoid
impacting performance.