Initial attempt at stack trace display.
This change adds a bit to the State object to indicate that a stack trace
exists at that location.  This is currently unused because maintaining this
bit slows down the loader.

Pulled out another class in DisplayLayout to abstract source data in
anticipation of using this for mmap and stack info.