Allow more than 36 bits of address space.
For Pin, it's impossible to directly limit the number of bits of address
space that will be used by the program. So we need a way in MemoryState to
represent the full 64 bits of address space without actually mapping all 64
bits of memory (which is impossible with mmap). To do this, add a linked
list of SparseArray objects that maps from the top set of bits to a
SparseArray. The StateArray stored in the link is the same as previous.
Update iterators and pages to correctly report addresses when using the
linked list.

This change slows down updateAddress() by roughly 25% for the test
program.